+++ docs/changelog.md
@@ -9,6 +9,7 @@
 - Added support for decision records.
 - Moved PlantUML support to a separate Structurizr.PlantUML project.
 - Separated the API client from the core library, and created a new Structurizr.Client project.
+- Renamed the "type" property on Section to "title".
 
 ## 0.7.2